fix(fork): a namespaced name's own ancestry (KEN-573) [no-changelog]
A namespaced fork name is the one destination whose name spells a path:
every render target folds the separator into a single leaf, but the local
source stores plugin/item nested. The slot being free said nothing about
what stood above it, so a fork named after an existing package landed
inside that package's tree and rode along in every render of it, and one
named through a symlinked plugin directory landed outside the source
entirely.

The slot is now asked of the catalog reader the rest of the engine
resolves this source with: SealedSource containment for the path, and
find_item for whether the plugin half is a package of its own. Nothing
here re-spells the local source's layout.

The name-vacancy refusals move to their own test module, matching the
source module they exercise.
